:construction: Migrate share plugin to new Flutter plugin API (and a new class)
Also doing some minor refactoring of the share code
